fix issue #6 compilation warnings
