+++ Sources/GraphQL/Map/Map.swift
@@ -663,11 +663,11 @@ extension Map : ExpressibleByStringLiteral {
     public init(unicodeScalarLiteral value: String) {
         self = .string(value)
     }
-
+    
     public init(extendedGraphemeClusterLiteral value: String) {
         self = .string(value)
     }
-
+    
     public init(stringLiteral value: StringLiteralType) {
         self = .string(value)
     }
@@ -682,133 +682,11 @@ extension Map : ExpressibleByArrayLiteral {
 extension Map : ExpressibleByDictionaryLiteral {
     public init(dictionaryLiteral elements: (String, Map)...) {
         var dictionary = [String: Map](minimumCapacity: elements.count)
-
+        
         for (key, value) in elements {
             dictionary[key] = value
         }
-
+        
         self = .dictionary(dictionary)
-    }
-}
-
-// MARK: CustomStringConvertible
-
-extension Map : CustomStringConvertible {
-    public var description: String {
-        var indentLevel = 0
-
-        let escapeMapping: [Character: String] = [
-            "\r": "\\r",
-            "\n": "\\n",
-            "\t": "\\t",
-            "\\": "\\\\",
-            "\"": "\\\"",
-
-            "\u{2028}": "\\u2028",
-            "\u{2029}": "\\u2029",
-
-            "\r\n": "\\r\\n"
-        ]
-
-        func escape(_ source: String) -> String {
-            var string = "\""
-
-            for character in source.characters {
-                if let escapedSymbol = escapeMapping[character] {
-                    string.append(escapedSymbol)
-                } else {
-                    string.append(character)
-                }
-            }
-
-            string.append("\"")
-            return string
-        }
-
-        func serialize(map: Map) -> String {
-            switch map {
-            case .null: return "null"
-            case .bool(let bool): return String(bool)
-            case .double(let number): return String(number)
-            case .int(let number): return String(number)
-            case .string(let string): return escape(string)
-            case .array(let array): return serialize(array: array)
-            case .dictionary(let dictionary): return serialize(dictionary: dictionary)
-            }
-        }
-        
-        func serialize(array: [Map]) -> String {
-            var string = "["
-
-            if _isDebugAssertConfiguration() {
-                indentLevel += 1
-            }
-            
-            for index in 0 ..< array.count {
-                if _isDebugAssertConfiguration() {
-                    string += "\n"
-                    string += indent()
-                }
-
-                string += serialize(map: array[index])
-                
-                if index != array.count - 1 {
-                    if _isDebugAssertConfiguration() {
-                        string += ", "
-                    } else {
-                        string += ","
-                    }
-                }
-            }
-
-            if _isDebugAssertConfiguration() {
-                indentLevel -= 1
-                return string + "\n" + indent() + "]"
-            } else {
-                return string + "]"
-            }
-        }
-        
-        func serialize(dictionary: [String: Map]) -> String {
-            var string = "{"
-            var index = 0
-
-            if _isDebugAssertConfiguration() {
-                indentLevel += 1
-            }
-            
-            for (key, value) in dictionary.sorted(by: {$0.0 < $1.0}) {
-                if _isDebugAssertConfiguration() {
-                    string += "\n"
-                    string += indent()
-                    string += escape(key) + ": " + serialize(map: value)
-                } else {
-                    string += escape(key) + ":" + serialize(map: value)
-                }
-
-                if index != dictionary.count - 1 {
-                    if _isDebugAssertConfiguration() {
-                        string += ", "
-                    } else {
-                        string += ","
-                    }
-                }
-                
-                index += 1
-            }
-
-            if _isDebugAssertConfiguration() {
-                indentLevel -= 1
-                return string + "\n" + indent() + "}"
-            } else {
-                return string + "}"
-            }
-        }
-
-        func indent() -> String {
-            return String(repeating: "    ", count: indentLevel)
-        }
-        
-        return serialize(map: self)
     }
 }
