+++ Tests/GraphQLTests/MapTests/MapTests.swift
@@ -548,7 +548,7 @@ public class MapTests : XCTestCase {
         let description = "{\"array\":[[],true,{},4.2,1969,null,\"foo\\nbar\"],\"bool\":true,\"dictionary\":{\"array\":[],\"bool\":true,\"dictionary\":{},\"double\":4.2,\"int\":1969,\"null\":null,\"string\":\"foo\\nbar\"},\"double\":4.2,\"int\":1969,\"null\":null,\"string\":\"foo\\nbar\"}"
 
         if _isDebugAssertConfiguration() {
-            XCTAssertEqual(buffer.description.characters.count, 465)
+            XCTAssertEqual(buffer.debugDescription.characters.count, 465)
         } else {
             XCTAssertEqual(buffer.description, description)
         }
